fix: resolve CodeQL Java code-scanning alerts at their true sources
Triaged 30 open CodeQL alerts on main via the SARIF code-flows: the
reported sink locations collapse to four root causes.

- insecure-randomness (24 alerts): RandFunction returned Math.random(),
  which CodeQL follows through config-string resolution into security
  sinks. Switched to a per-thread SecureRandom (secure source, no
  cross-thread lock contention on the query hot path).
- tainted-format-string (4 alerts): BinarySerializer concatenated an
  untrusted value into a log message; pass it as a %s argument instead.
- path-injection (1 alert): untrusted Neo4j import labels became schema
  type names and on-disk bucket file names. Validate labels at the
  import boundary, rejecting path separators and parent traversal.
- implicit-cast-in-compound-assignment (1 alert): widen totalRecords to
  long in CountStepTest.

Added Neo4jImporterIT regression tests for malicious node/edge labels.

+++ engine/src/test/java/com/arcadedb/query/sql/executor/CountStepTest.java
@@ -78,7 +78,7 @@ class CountStepTest extends TestHelper {
     });
 
     final ResultSet result = database.query("sql", "SELECT status, count(*) as total FROM Order GROUP BY status");
-    int totalRecords = 0;
+    long totalRecords = 0;
     while (result.hasNext()) {
       final Result item = result.next();
       final long count = item.<Number>getProperty("total").longValue();
